requirements.txt
asgiref==3.8.1
Django==5.1.5
django-autoslug==1.9.9
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
Markdown==3.7
packaging==24.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
segno~=1.6.1
python-dotenv~=1.0.1
future~=1.0.0